requirements.txt
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.2
mitmproxy>=10.2.0
bcrypt==4.0.1
httpx>=0.25.0
redis>=5.0.0
pydantic-settings>=2.1.0
loguru>=0.7.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.0
python-multipart>=0.0.6
